requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==24.1.1
packaging==26.0
pillow==12.1.0
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
